code/requirements_public_train.txt
# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
# SPDX-License-Identifier: LicenseRef-NvidiaProprietary
#
# NVIDIA CORPORATION, its affiliates and licensors retain all intellectual
# property and proprietary rights in and to this material, related
# documentation and any modifications thereto. Any use, reproduction,
# disclosure or distribution of this material and related documentation
# without an express license agreement from NVIDIA CORPORATION or
# its affiliates is strictly prohibited.

# Public training requirements (Torch-only pipeline).
# Install a PyTorch build compatible with your Python/CUDA stack.
# Target Python versions: 3.11, 3.12, 3.13
-r requirements_public_inference.txt
tensorboard
torchinfo
